Better handle stripping CVSROOTs from CVS paths.
We had code to strip the absolute path prefixes from CVS paths in diffs,
but it didn't handle CVSROOTs that didn't separate the domain and path
with a ':'. We now handle this case a bit better by finding where that
path begins.

Also fixes a bug where the attic filename was being generated incorrectly
if the file didn't contain a '/' in the name.

Reviewed at http://reviews.reviewboard.org/r/1684/