Don't enable extensions in manage.py for unit tests or runserver.
If enabling extensions in manage.py when running unit tests or runserver
(the devserver), we end up enabling extensions twice in a way that causes
breakages. We now prevent the loading of extensions under this case. The
proper loading will happen when the first URL is fetched and Review Board
is initialized.

Patch by Mike Conley.
Reviewed at http://reviews.reviewboard.org/r/1679/