Don't use "/Attic" as the Attic path in CVS.
If a diff file in CVS doesn't contain any path indicators at all, the
default CVS Attic path would begin with "/Attic". It turns out that this
is less than useful, as CVS will just complain about the fact that it's an
absolute path, and fail. We now just skip testing the attic path in this
case.

This fixes the nightly builds.