Improve move detection.
This change greatly improves move detection by finding moved ranges within
insert/delete chunks. Previously, move detection only worked when entire
insert/delete chunks matched, but now if only a few lines within a chunk appear
to have moved, it'll detect that.

It also filters out some things that used to appear as moves but were pretty
useless. Mainly, blank or whitespace-only lines, and lines containing
non-alphanumeric data only (such as "}", "/*", " *", "*/", and other such
things). These are filtered only if the move range consists *only* of such
things. If some real content exists within the move range, the entire thing is
preserved.

We also filter out moves that have more than one destination. If a particular
block was deleted from one section and then appears to be inserted in two or
more places, then it hasn't really moved, and it's likely something that is not
particularly interesting. We may decide later to list all the destination areas
instead, but for now it should be fine.

The new algorithm is a little more complicated, but it's been fully documented.
Due to the more complex checking, it is in theory a little slower, but only at
the worst-case scenarios of many identical insert/delete ranges.

Reviewed at http://reviews.reviewboard.org/r/1683/