Fix several problems reported by PyFlakes.
PyFlakes was reporting a lot of undefined and unused variables. The unused
variables/functions were particularly a problem, though rarely hit. Mostly
error conditions from the new web API code.

All PyFlakes warnings are now fixed, and the few legitimate code paths that
PyFlakes was tripping on are now covered in our pyflakes.exclude file.

There is one change in particular I should note for future reference. There
was a code path in myersdiff.py that assigned to an incorrectly named variable,
resulting in that value never being returned correctly from the function. It's
possible some bugs have resulted from this in the past, but it's hard to tell.
I've fixed the variable name, and we'll just have to wait and see if anything
changes.