Fix a rendering bug in the diff viewer when using Pygments.
By default, Pygments generates a series of wrappers around syntax-highlighted
content. These were wrapped using <div> and <pre>, with the assumption on
the Pygments end being that they were intended to be directly inserted into
an HTML element and not manipulated in the way we manipulate them.

The visual effect was that the first line of the diff viewer would sometimes
appear with a double height.

This change strips out the wrappers so that we don't end up with a slightly
broken diff viewer.

Patch by Eduardo Felipe Castegnaro.
Reviewed at http://reviews.reviewboard.org/r/1698/