Move the webapi 1.0 docs to webapi/1.0/.
This is the beginning of the documentation process for the new webapi. The
old docs are moving to a 1.0/ directory, and the new docs will be in a 2.0/.
The toctree has been set up for the new docs.

This also removes some of the information from the 1.0 overview that documents
things that would have been part of the API but instead have moved to the new
API.