Properly handle links to comment resources.
Comment resources could live in two main places: On reviews, or on replies to
reviews. We were only showing the links to the ones on the reviews, ignoring
the reply ones. This broke our nice linking. We now return the proper
resource depending on where the particular comment should live. This applies
to both diff and screenshot resources.

As part of this, to make the naming more clear, the plain "*CommentResource"
classes have been renamed to "*DiffCommentResource".