Added the new webapi 2.0 docs.
This adds back the main index for the webapi docs, and fills in the webapi 2.0
overview, authentication, and error listings.

For the resource docs, we use the new introspection support in Djblets and
Review Board. The new docs just simply make use of the new role and passes the
right parameters to it. The result is that documentation is generated from
the capabilities and internal docs of the resource classes, which will help
to keep them maintained.