Attempt to fix pass-through to the built-in auth backend when using AD.
There's a bug with ActiveDirectory support where, if an account couldn't
be found, it would fail to fall back to the built-in auth. It appears we're
not giving up soon enough when searching ActiveDirectory for a user fails.
We now check if the lookup returned nothing, and then bail.

This isn't a verified fix for the users that hit the problem, but does appear
to be the right thing to do regardless.

Hopefully fixes bug #1536.