Show SCM errors (such as File Not Found) in the New Review Request page.
The code to display errors in the New Review Request page regressed, and
wasn't showing errors for File Not Found or "The Base Diff Path field is
required." We weren't catching all the right exception types. This led to
a 500 Internal Server Error. Now we display the right errors, and also
handle other SCM-related errors.