Allow @valid_prefs_required to be used with @check_login_required.
The review request details page requires a profile for authenticated users,
so @valid_prefs_required was recently added for that view. However, this broke
anonymous users, because @valid_prefs_required assumes that the user could only
reach that point if they were logged in. This wasn't the case here, so now
@valid_prefs_required checks to make sure the user is authenticated before
performing the check. If the user is not authenticated, it assumes that the
view will only check the profile for authenticated users, and passes control
on to it.