Improve usability of screenshot drag-and-drop.
Screenshot drag-and-drop failed to properly scroll to the drop area, which on
a brand new review request made it very hard to drop a screenshot. We now
ensure that after animation, we scroll into view.

This also prevents a JavaScript error when the drag events get a bit out of
sync, which appears to be able to happen on Firefox when dragging into the
container and then out via chrome rather than the page. We no longer assume
that the dataTransfer object is non-null.