Support PDF/PS output for the Admin Guide and User Manual.
To generate PDF/PS manuals, we need to support generating Latex files,
which had a number of problems. Our web API guides use nested tables and
other markup that fails currently with the Latex writer. We also had
a badly organized Table of Contents, which was fine for the HTML output but
resulted in many duplicated pages for Latex.

We now output two separate files, one for the Admin Guide and one for the
User Manual. We don't output the API guide at all. We also use a real
visible toctree instead of some toctrees reaching into sub-nodes with matching
hidden toctrees, which was confusing Latex. The result is a difference in
how the manual looks in HTML, but with a proper Table of Contents throughout.