Add a toggle to hide / show trailing or extra whitespaces
Patch from Laurent Nicolas. Our users could not decide whether they wanted extra
whitespaces to be highlighted or not.  Some people love it, others find it too
distracting, the more for legacy code.  This change provides a button, modeled
on the collapse change button.  The choice is also saved in a cookie, so that
each user can have her own preference.

This is different from Show/Hide Whitespace changes, and the 2 options cohabit
well, I verified that after upgrading to 1.5 beta1.  (I had to change gt(0) to
gt(1) in the javascript.  I would rather use a class name associated to the li
tag, but I was concerned about a potential performance impact.)

I was also wondering if I should make it more dynamic, ala Show/Hide Whitespace
changes, but I see it as a more static choice, depending on the code you're
working on.

Testing done: tried it out.