Correctly mark revision as UNKNOWN for mercurial mqueue patches
Patch from Eduardo Felipe Castegnaro. Patches that come from mercurial mqueues
do not have version information on them, so the revision parameter comes as an
empty string, when it should be marked an treated as UNKNOWN.

This fixes an issue that ReviewBoard would always fail to accept a patch from an
mqueue, saying that the repository didn't contain the file, even when it did.