Stop shipping our own copy of the Django Admin media files.
For historical reasons, we've shipped the Django Admin media files as part
of our build directory. This comes from a time before rb-site existed, when
we needed a single media directory with everything inside it. However, that's
just problematic these days.

Now we let the Django devserver handle the /media/admin path as it prefers
to, and we link the proper path during rb-site upgrade. This lets us reduce
our package size and prevents problems people may hit with different
versions of Django.