Don't hide the screenshots area when uploading a new screenshot.
Uploading a new screenshot through the dialog would cause the screenshot
area to disappear. This is due to calling empty() to determine if the
area was empty. However, this doesn't check for emptiness; rather it actually
empties the area. The screenshots reappear on reload, but it can be confusing.

We now check the length of the jQuery element, as we should have originally.