Show permission denied errors when accessing unreachable Git repositories.
When adding a Git repository pointing to a directory that the web server
doesn't have permissions to browse, we would hit an error saying that
information couldn't be retrieved. We now check after failure if we even
have read permissions on the directory and display an appropriate error
message.

Fixes bug #1765.