Review draft, invalid user/group warning
This is addressing the issue here:
http://code.google.com/p/reviewboard/issues/detail?id=1773

When creating a review draft, if a user or group doesn't exist, a message will
notify the user of the error.  Currently the invalid user or group would simply
be removed.

Message varies based on number of mistakes.

Testing done: Manual testing.  Tried editing everything on the draft page, seems
to be working fine.