Collapsible reviews.
In order to tidy existing and future review requests I have implemented
collapsible reviews.  There are two criteria for the starting state of a review.

a) The review is collapsed if it is older than the last change to the review
request.

b) The review is collapsed if all of its replies are older than the last time
the user visited the page.

I wanted to use cookies to allow the site to remember which reviews had been
collapsed or expanded, but couldn't figure out a good way of implementing it.
Hopefully I can make this available later.