Remove the LocalSite.nextid DeleteField evolution.
It turns out that Django Evolution fails when performing a single database
evolution that includes both the creation of a model and a deletion of a
field on that model. The reason is that the initial creation of the model
won't reference the field, so Django Evolution won't find the field to
delete.

This is only a problem if doing both evolutions at the same time. If the
model with the field is first added and the schema is stored, and *then*
deleted, it's safe. However, in our case, anybody upgrading from RB 1.5
to 1.6 will encounter this problem.

Since this is a hard thing for Django Evolution to know, the easiest fix is
just to remove the evolution from Review Board. Any developer working
day-to-day on hte codebase can do their own 'evolve --hint --execute' to
handle the deletion of that particular field, but likely this will only be
a couple of people.