Filter repositories based on a user's access.
This provides a Repository.objects.accessible() query function that allows
for filtering based on a user's access to the repository. It can also optionally
show hidden (but accessible) repositories if visible_only is set to False.

All call sites have been updated to use the new accessible() function and to
check if a repository is accessible when accessing it directly. This includes
the New Review Request page.

Reviewed at http://reviews.reviewboard.org/r/1915/