Fix issue 1870: base path does not work with svn-hook-postcommit-review
I realize this is a contributed script and probably isn't all that supported but
the base path parameter in /contrib/tools/svn-hook-postcommit-review is not
functioning.

I get an error from post-review showing that it included the quotes in the paths
it passed to svn diff and other commands (forgot to copy/paste).

Testing done: Tested on centos 5 64bit with svn 1.6.12, mod dav svn, reviewboard
1.5.
Reviewed at http://reviews.reviewboard.org/r/1875/