Fix evolve/syncdb commands with extensions when coming from an older database.
manage.py on the extensions branch assumed the existance of the proper columns
in the extensions tables, and that the tables even existed. This prevented
syncdb or evolve from running when upgrading from a non-extension-enabled
database.

We now look for a DatabaseError on extension manager loading and skip the
extension loading if we see it.