Always return the same result when doing a PUT on a ReviewRequestDraftResource.
Previously, when PUTing public=1 on a ReviewRequestDraftResource, we would
get a 303 SEE OTHER redirect response, pointing to the draft. While useful,
this made any interaction with AJAX libraries impossible, as the browser
hides the redirect, disallowing any custom handling of it, and doesn't
even preserve the HTTP Accept header, leading to parsing errors.

Now, we just return the same result that other updates to that resource
return, which is an indication of which key was modified and which value
it now contains.