Fix several issues with scoping and XML output from upload forms.
JavaScript scoping bit us, causing the expected 'options' variable to
end up being overridden when used with the new ready() functions. We now
limit the scope using apply(this, arguments).

We also were receiving XML output from any forms during uploads. The
reason is that we relied upon the browser to perform the request, but
it requests HTML/XML output, whereas we needed JSON. We now hard-code
"api_format=json" in the POST data in this case.