Add LocalSite rooting and unit tests for all of the rest of the REST API.
This change adds the namespacing and unit tests for all the different comments
and reply resources as well as a few others. I've also updated a bunch of
resources to base their URLs off the review_request_resource's, in order to do
the right thing with local_id (many of these tests wouldn't work without that).