Add API for change number-related updates.
This adds some missing API for both updating the associated change number
for a review request, and updating the review request's information based on
the change number. These two things were missing from the new API, and were
provided by the old one.

To simply update the review request's information based on a change number,
a PUT changenum=<changenum> with the existing change number should be called.
Specifying a different change number will attempt to update which change the
review request is pointing to, and then update the information from that.

These are needed for post-review, and will be backported to release-1.5.x.