Add capability lookup for auth backends and fix regressions
This is the second part to the big authentication change.

We no longer check for specific auth backends when determining what data can be changed or which links to show. Now we ask the auth backends themselves to find out what is supported.

Standard auth now goes through a new StandardAuthBackend class, which provides the capabilities we care about and forwards any requests down to Django's standard auth module. As a result, we can now use a form instead of stuffing that data into AuthenticationSettingsForm

This also fixes some problems with loading custom auth backends, as the
siteconfig code was ignoring them on load.