Add documentation on writing new-style authentication modules.
This adds new documentation for writing the new authentication modules. This
includes a definition of all the new attributes and the methods, how to write
settings forms, accessing settings, and packaging the authentication module.

The old documentation has been moved to legacy-auth-backends.txt. Both the new
and old docs reference each other.