Add support for updating User information on the backend.
This provides new AuthBackend functions for updating the password,
e-mail address, and full name on the backend.

For the e-mail address and full name, we first store in the User object
and then allow the auth backend to update the backend based on the values in
the User object.

For passwords, we actually have the auth backend handle all aspects of
the updating. This is because we don't necessarily want to store data in
the User object. The StandardAuthBackend contains the logic for updating
the User object, but other backends likely won't want to modify that, in
order for lookups on the server to work.

Reviewed at http://reviews.reviewboard.org/r/2000/