Add a ReviewRequest.get_display_id to generalize the ID field.
In a lot of places we had to check both id and local_id in order to do something
interesting for the user (links, display on the template, etc). This change
generalizes that with a get_display_id() method, and fixes the ID printed in the
review request box and trophy calculation.