Add support for URL reversing for local sites.
This adds a new local_site_reverse function in reviewboard.site.urlresolvers
that can, given a local_site_name or request, attempt to reverse the URL
with that site name. This makes it easy to write a function that points
to the correct URL depending on whether or not the user is accessing a local
site.

Call sites have been updated to use the new function. The review request
models have also been updated to use this in the get_absolute_url functions.
There shouldn't be any manual URL generation left at this point.

Reviewed at http://reviews.reviewboard.org/r/2008/