Fix a problem with SSH host checking on Git repositories.
SSH host checking on Git repositories was busted when using the
user@address:/path form, without a leading ssh://. We were using the
raw provided path, and not the normalized path, when doing the verification.
Since it didn't appear as a SSH path, the host wasn't checked.

This fixes one issue with checking the host for private Git repositories on
GitHub.