Fix a SSH URL parsing issue breaking private GitHub repositories.
The code to parse a SSH-based URL to extract the hostname and username
was giving precedence to the username provided to the function, rather than
that in the URL. In the case of private GitHub repositories, the server
needed to be accessed with the "git" user, but we were using the
repository owner's username instead.

We now give precedence to the URL, using the provided username as the
fallback. This fixes adding private GitHub repositories (so long as
the Review Board instance has a DSA or RSA keypair and the public key is
added to the GitHub repository).