Fix hosting service matching for private GitHub repositories.
The code that attempts to parse a URL and find the matching hosting
service didn't work with private GitHub repositories. It was too
strict and required a mirror path, which we don't use for those. We only
really care about the Repository instance and the path, so check for those.