Add a settings page for SSH key information/generation.
This introduces a page for viewing/creating a SSH key for the user
running Review Board. This makes it easier to get up and going with
SSH-backed repositories without needing access to the terminal.

If an SSH key (id_dsa or id_rsa file) exists in the .ssh directory,
information on the key will be displayed. Specifically, the key type,
how many bits, the fingerprint, and the public key.

If the SSH key does not exist, a button will be shown that, when clicked,
generates and saves a key file. The page will be updated to instantly show
the new key.

Reviewed at http://reviews.reviewboard.org/r/2015/