Fix authentication with the web API on local sites.
The @webapi_check_local_site decorator was unconditionally returning
Permission Denied when the user didn't have permission on the site.
However, this check makes no sense for unauthenticated users. Now, like
the other uses of Permission Denied errors, we return Not Logged In when
the user hasn't logged in yet, allowing the client to attempt a login
and re-access of the API.