Don't attempt to unconditionally reverse the dashboard URL in user_preferences.
The user_preferences view was using the dashboard URL as the fallback if a
?next_page= URL wasn't specified. However, if it was specified, we would
still compute this, which was an extra unnecessary check and could break
if the non-localsite dashboard was disabled. We now only reverse the URL if
The next_page URL wasn't provided.

This doesn't yet prepare the URL reversing to account for local sites and to
go to the correct URL.