Fix autocomplete drop-downs for users and groups.
This change fixes two problems with the autocomplete drop-downs. First of all,
we were using the old API path, which worked fine except didn't do any
local-site filtering and won't work when we turn off the old API. Second, the
/api/users/ resource was using "query" instead of "search_q" in some of the LIKE
relations, which created a complex query that promptly raised a mysql
operational error.