Provide more helpful information on repository authentication failures.
When failing to authenticate with a repository, we provided a very
unhelpful error talking about the username and password being wrong.
However, sometimes the username/password aren't even allowed, and the
real factor is the public key. In those cases, the user wouldn't even
know.

Now we refrain from mentioning specifics, but instead list of authentication
types we know were tried. This of course assumes the backend is using this
new error, which right now is only used for SSH errors.

When we get an authentication error that mentions that a public key is
an acceptable form of authentication, and no public key has been
generated on the server, we tell the user this and give them a link (which
opens in a new tab/window) to generate a new key.

Reviewed at http://reviews.reviewboard.org/r/2019/