Add support for uploading SSH keys to a server.
This introduces the abliity to upload an existing id_dsa or id_rsa key
to the server. The key will be validated and then stored as id_dsa or id_rsa,
depending on the type (and regardless of the original filename).

This allows an administrator to easily provide a known good key instead of
having to copy the file over as root.

Reviewed at http://reviews.reviewboard.org/r/2024/